Use long lines in the .txt file
The .txt file works in GitHub issues, but the markdown variant
used by GitHub issues forces line returns when the text file does
(which isn't true for "normal" markdown). Thus, text files with lots
of line break look funny in it.

This commit merges the lines of a parapgraph into a single long line.
This works in both GitHub issues markdown and in email.
The goal is to have a single form we can use everywhere.

This also removes the intro comment about needing to handle validation codes;
our text now *does* handle validation codes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>