Add Georg Kunz and organizational affiliations
Add Georg Kunz as a notifier in the README.

Also, add organizational affiliations as well as a note that
"Organizational affiliations are *only* shown to clarify who we mean."

One challenge for notifiers is convincing projects that they aren't
spammers or scammers. Listing them on the repo README, along with
their organizations, will hopefully reduce that problem.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>