Add account ids of notifiers
GitHub @balloob reported that
"Home Assistant received this issue" home-assistant/core#61455
but wasn't sure if it was legit. We listed notfier *names* but
@balloob correctly noted that it's easy to set names on another account.

This commit adds GitHub (and at one GitLab) account names, where I know them,
to make it easier for many projects to confirm if a given notification really
is a message from this project.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>